Conflicts:
      application/apps/indexer/Cargo.lock
      application/apps/indexer/sources/src/binary/raw.rs
      application/apps/rustcore/rs-bindings/Cargo.lock
      cli/development-cli/CHANGELOG.md
      cli/development-cli/Cargo.lock
      cli/development-cli/Cargo.toml

+++ application/apps/indexer/sources/src/command/process.rs
@@ -1,5 +1,5 @@
 use crate::{ByteSource, Error as SourceError, ReloadInfo, SourceFilter};
-use buf_redux::Buffer;
+use bufread::DeqBuffer;
 use regex::{Captures, Regex};
 use shellexpand::tilde;
 use std::{collections::HashMap, ffi::OsString, path::PathBuf, process::Stdio};
@@ -30,7 +30,7 @@ pub enum ProcessError {
 
 pub struct ProcessSource {
     process: Child,
-    buffer: Buffer,
+    buffer: DeqBuffer,
     stdout: FramedRead<ChildStdout, LinesCodec>,
     stderr: FramedRead<ChildStderr, LinesCodec>,
     stdin: ChildStdin,
@@ -38,8 +38,11 @@ pub struct ProcessSource {
 
 impl Drop for ProcessSource {
     fn drop(&mut self) {
-        if let Err(err) = self.process.start_kill() {
-            error!("Fail to kill child process: {err}");
+        let is_process_alive = self.process.try_wait().is_ok_and(|state| state.is_none());
+        if is_process_alive {
+            if let Err(err) = self.process.start_kill() {
+                warn!("Fail to kill child process: {err}");
+            }
         }
     }
 }
@@ -144,7 +147,7 @@ impl ProcessSource {
             .ok_or_else(|| ProcessError::Setup(String::from("Fail to get stdin handle")))?;
         Ok(Self {
             process,
-            buffer: Buffer::new(),
+            buffer: DeqBuffer::new(8192),
             stdout,
             stderr,
             stdin,
@@ -183,9 +186,9 @@ impl ByteSource for ProcessSource {
         }
         if let Some(Ok(line)) = output {
             let stored = line.len() + 1;
-            self.buffer.copy_from_slice(line.as_bytes());
-            self.buffer.copy_from_slice(b"\n");
-            let available_bytes = self.buffer.len();
+            self.buffer.write_from(line.as_bytes());
+            self.buffer.write_from(b"\n");
+            let available_bytes = self.buffer.read_available();
             Ok(Some(ReloadInfo::new(stored, available_bytes, 0, None)))
         } else if let Some(Err(err)) = output {
             Err(SourceError::Unrecoverable(format!("{err}")))
@@ -195,15 +198,15 @@ impl ByteSource for ProcessSource {
     }
 
     fn current_slice(&self) -> &[u8] {
-        self.buffer.buf()
+        self.buffer.read_slice()
     }
 
     fn consume(&mut self, offset: usize) {
-        self.buffer.consume(offset);
+        self.buffer.read_done(offset);
     }
 
     fn len(&self) -> usize {
-        self.buffer.len()
+        self.buffer.read_available()
     }
 
     fn is_empty(&self) -> bool {
